% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyEnhancedDimPlot.R
\name{emptyEnhancedDimPlot}
\alias{emptyEnhancedDimPlot}
\alias{emptyEnhancedDimPlot.Seurat}
\title{emptyEnhancedDimPlot}
\usage{
emptyEnhancedDimPlot(object, ...)

\method{emptyEnhancedDimPlot}{Seurat}(object, reduction = "tsne",
  grouping_var = NULL, group_plot = NULL, dim_1 = 1, dim_2 = 2,
  pt_size = 1, alpha = 1, force = 1, label = TRUE,
  label_size = 3, label_text_color = "black", ...)
}
\arguments{
\item{object}{scRNAseq data object}

\item{...}{Additional parameters}

\item{reduction}{dimensional reduction to use. Default: tsne}

\item{grouping_var}{Metadata variable to use in grouping data. If none is provided,
the current ident will be used.  Default: NULL}

\item{group_plot}{If provided, only this identity group will be displayed. Default: NULL}

\item{dim_1}{Dimension to display along the x-axis. Default: 1}

\item{dim_2}{Dimension to display along the y-axis. Default: 2}

\item{pt_size}{Point size. Default: 1}

\item{alpha}{Point alpha. Default: 1 (i.e. fully opaque)}

\item{force}{Force parameter to pass to ggrepel.  See \code{\link{geom_label_repel}}}

\item{label}{Should labels be shown? Default: TRUE}

\item{label_size}{Label font size. Default: 3}

\item{label_text_color}{Label font color. Default: black}
}
\description{
Create a scatter plot where the points are not displayed but labels
for each group are.
}
\examples{

}
