% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancedFeaturePlot.R
\name{enhancedFeaturePlot}
\alias{enhancedFeaturePlot}
\alias{enhancedFeaturePlot.Seurat}
\title{enhancedFeaturePlot}
\usage{
enhancedFeaturePlot(object, ...)

\method{enhancedFeaturePlot}{Seurat}(object, feature, reduction = "tsne",
  grouping_var = NULL, subset_labels = NULL, group_plot = NULL,
  split_by = NULL, lower.cutoff = 0.05, upper.cutoff = 0.95,
  dim_1 = 1, dim_2 = 2, pt_size = 1, alpha = 0.5, force = 1,
  label = TRUE, label_size = 3, label_text_color = "black",
  scale_colors = c("grey90", "red"), fill_palette = NULL, ...)
}
\arguments{
\item{object}{scRNAseq data object}

\item{...}{Additional parameters}

\item{feature}{Feature to use for coloring.  Anything FetchData can extract.}

\item{reduction}{dimensional reduction to use. Default: tsne}

\item{grouping_var}{Metadata variable to use in grouping data. If none is provided,
the current ident will be used.  Default: NULL}

\item{subset_labels}{A list of the groups in \code{grouping_var} to display.  By default,
all are displayed.  Default: NULL}

\item{group_plot}{If provided, only this identity group will be displayed. Default: NULL}

\item{split_by}{Metadata variable to use in faceting the plots. Default: NULL}

\item{lower.cutoff}{Remove expression values below this quantile. Expressed as decimal. Default: 0.05}

\item{upper.cutoff}{Remove expression values above this quantile. Expressed as decimal. Default: 0.95}

\item{dim_1}{Dimension to display along the x-axis. Default: 1}

\item{dim_2}{Dimension to display along the y-axis. Default: 2}

\item{pt_size}{Point size. Default: 1}

\item{alpha}{Point alpha. Default: 1 (i.e. fully opaque)}

\item{force}{Force parameter to pass to ggrepel.  See \code{\link{geom_label_repel}}}

\item{label}{Should labels be shown? Default: TRUE}

\item{label_size}{Label font size. Default: 3}

\item{label_text_color}{Label font color. Default: black}

\item{scale_colors}{Colors to use for expression gradient.
Takes a list of \code{c(low, high)}.  Default: c("grey90","red")}

\item{fill_palette}{Palette to use for point fill.  By default, this matches \code{color_palette}}
}
\description{
A customized version of Seurat's DimPlot. Used for
plotting dimensional reductions.
}
\examples{

}
