% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighlightGroupDimPlot.R
\name{HighlightGroupDimPlot}
\alias{HighlightGroupDimPlot}
\alias{HighlightGroupDimPlot.Seurat}
\title{HighlightGroupDimPlot}
\usage{
HighlightGroupDimPlot(object, ...)

\method{HighlightGroupDimPlot}{Seurat}(object, reduction = "tsne",
  group_by = NULL, highlight_group = NULL,
  highlight_color = "#E41A1C", contrast_color = "#AAAAAA", dim_1 = 1,
  dim_2 = 2, pt_size = 1, highlight_alpha = 1,
  contrast_alpha = 0.75, force = 1, label = TRUE, label_size = 3,
  label_text_color = "black")
}
\arguments{
\item{object}{scRNAseq data object}

\item{reduction}{dimensional reduction to use. Default: tsne}

\item{group_by}{Metadata variable to use in grouping data. If none is provided,
the current ident will be used.  Default: NULL}

\item{highlight_group}{Particular group to highlight. Default: NULL}

\item{highlight_color}{Color to highlight chosen group with. Default: red}

\item{contrast_color}{Color to use for all other groups. Default: grey}

\item{dim_1}{Dimension to display along the x-axis. Default: 1}

\item{dim_2}{Dimension to display along the y-axis. Default: 2}

\item{pt_size}{Point size. Default: 1}

\item{force}{Force parameter to pass to ggrepel.  See \code{ggrepel::\link[geom_label_repel]{geom_label_repel}}}

\item{label}{Should labels be shown? Default: TRUE}

\item{label_size}{Label font size. Default: 3}

\item{label_text_color}{Label font color. Default: black}

\item{alpha}{Point alpha. Default: 1 (i.e. fully opaque)}
}
\description{
HighlightGroupDimPlot
}
