% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepPalette.R
\name{PrepPalette}
\alias{PrepPalette}
\title{PrepPalette}
\usage{
PrepPalette(palette_use, bins)
}
\arguments{
\item{palette_use}{The name of a palette to use.  Must be a palette available in
the Paletteer package.  If there are more unique identities than colors in the
palette, additional values will be created by interpolation.}

\item{bins}{Number of discrete colors to generate from palette.}
}
\value{
A list containing color values.
}
\description{
Given a palette name and data frame with clustering or grouping
information in an 'ident' column, return a palette containing a color for each
unique cluster identity.
}
\examples{

}
